src/ragapp/pyproject.toml
[tool.poetry]
name = "rag-box"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.111.1"
python-dotenv = "^1.0.0"
uvicorn = { version = "^0.23.2", extras = ["standard"] }
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
aiostream = "^0.5.2"
chromadb = "0.5.1"
ollama = "^0.3.0"
e2b-code-interpreter = "0.0.11b38"
sqlmodel = "^0.0.21"
jsonschema = "^4.22.0"
cachetools = "^5.3.3"

# llama-index packages
llama-index = "0.11.17"
# llms
llama-index-llms-openai = "^0.2.0"
llama-index-llms-openai-like = "^0.2.0"
llama-index-llms-azure-openai = "^0.2.1"
llama-index-llms-gemini = "^0.3.4"
llama-index-llms-ollama = "^0.3.1"
llama-index-llms-mistralai = "^0.2.3"
llama-index-llms-groq = "^0.2.0"
# embeddings
llama-index-embeddings-openai = "^0.2.3"
llama-index-embeddings-azure-openai = "^0.2.5"
llama-index-embeddings-gemini = "^0.2.0"
llama-index-embeddings-mistralai = "^0.2.0"
llama-index-embeddings-fastembed = "^0.2.0"
llama-index-embeddings-ollama = "^0.3.0"
# vector stores
llama-index-vector-stores-chroma = "^0.2.0"
llama-index-vector-stores-qdrant = "^0.3.0"
# agent
llama-index-agent-openai = "^0.3.0"
# tools
llama-index-tools-wikipedia = "^0.2.0"
llama-index-tools-duckduckgo = "^0.2.0"
llama-index-tools-openapi = "^0.2.0"
llama-index-tools-requests = "^0.2.0"
# others
llama-index-indices-managed-llama-cloud = "^0.3.0"
llama-index-postprocessor-cohere-rerank = "^0.2.0"
markdown = "^3.7"
xhtml2pdf = "^0.2.16"



[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
